Delete limit from sync token
The Sync API has brand new 'limit' parameter. It can only be set when initial: true, and should not be sent with syncTokens. This deletes the limit from the sync query if a sync_token is set, otherwise you get the error: When passing a sync token you can not pass any other parameters. If you want to sync with a specific limit, start a new sync with "initial".