fix: ensure the node bundle uses native async/await
We bumped the minimum version of Node to v12.
This should have happened in the last release already, but was forgotten about.
While v10 is still on  Maintenance LTS, it will be EOL by the end of Q1 2021.
Thus we're jumping directly to v12 - however most likely the library will work perfectly fine also with v10 since currently we are not using any features unique to v12.