feat(taxonomy): add concepts to entry metadata [GROOT-1440] (#2304)
# Summary
Add Taxonomy concept links to entries (This feature i currently in beta).

# Description
To allow beta customers to use the newly exposed concepts field on metadata, we add it as an optional field.
For now, taxonomy concepts are exclusive to Entry. The change introduces the smallest possible footprint to introduce this variation for Entry metadata. It should also allow for easy adoption of the feature on assets at a later point in time.